+++ EZLogger/HostForms/PatientInfoHost.vb
@@ -4,8 +4,10 @@ Public Class PatientInfoHost
     Inherits Form
 
     Private Sub PatientInfoHost_Load(sender As Object, e As EventArgs) Handles MyBase.Load
-        Dim view As New EZLogger.Views.PatientInfoView()
+        Dim view As New EZLogger.Views.PatientInfoView(Me)
         ElementHost1.Dock = DockStyle.Fill
         ElementHost1.Child = view
+        Me.Text = ""
+        Me.ClientSize = New Drawing.Size(400, 755)
     End Sub
 End Class